requirements.txt
loguru
tqdm
pymongo
omegaconf
tweepy
joblib
numpy
xgboost
regex
requests
requests-oauthlib
bson
pandas
jsonlines
searchtweets-v2
slackclient